% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_datasets.R
\docType{data}
\name{cancer_dataset}
\alias{cancer_dataset}
\title{Wisconsin breast cancer data set}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 683 rows and 11 columns.
}
\usage{
cancer_dataset
}
\description{
Wisconsin breast cancer data set
}
\keyword{datasets}
