% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_datasets.R
\docType{data}
\name{chess_dataset}
\alias{chess_dataset}
\title{King-rook vs king-pawn data set}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3196 rows and 37 columns.
}
\usage{
chess_dataset
}
\description{
King-rook vs king-pawn data set
}
\keyword{datasets}
