% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kohavi.R
\name{cv_experiment}
\alias{cv_experiment}
\title{Estimate the accuracy of models on a dataset using cross-validation}
\usage{
cv_experiment(
  dataset_name,
  dataset,
  nsamples,
  times = 50,
  fold_counts = my_fold_counts,
  models = my_model_list,
  seed = 2021
)
}
\arguments{
\item{dataset_name}{a unique string identifying the dataset}

\item{dataset}{the data set to use}

\item{nsamples}{how large the training set should be}

\item{times}{how many times to train and test the models, default 500}

\item{fold_counts}{folds to use, with a negative number meaning leave-k-out}

\item{models}{the models to train}

\item{seed}{random seed}
}
\value{
a matrix of accuracies, where each row corresponds to a model
}
\description{
\code{cv_experiment} tests the effectiveness of cross-validation for
estimating the accuracy of different models.
}
