% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kohavi.R
\name{cv_lr_experiment}
\alias{cv_lr_experiment}
\title{Repeat the CV experiment for logistic regression with different penalties}
\usage{
cv_lr_experiment(
  dataset_name,
  dataset,
  nsamples,
  folds,
  times = 50,
  penalties = my_penalty_list,
  seed = 2021
)
}
\arguments{
\item{dataset_name}{a unique string identifying the dataset}

\item{dataset}{the data set to use}

\item{nsamples}{how large the training set should be}

\item{folds}{number of folds to use, with a -1 meaning leave-one-out}

\item{times}{how many times to train and test the models, default 500}

\item{seed}{random seed}
}
\value{
a matrix of accuracies, where each row corresponds to a model
}
\description{
\code{cv_lr_experiment}.
}
