% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_datasets.R
\docType{data}
\name{hypothyroid_dataset}
\alias{hypothyroid_dataset}
\title{Thyroid disease data set}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3163 rows and 26 columns.
}
\usage{
hypothyroid_dataset
}
\description{
Thyroid disease data set
}
\keyword{datasets}
