% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.R
\name{inversions}
\alias{inversions}
\title{Count order inversions}
\usage{
inversions(xs, decreasing = FALSE)
}
\description{
This function uses an insertion sort to count the number of transpositions
necessary to sort the list \code{xs}.
}
