% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_datasets.R
\docType{data}
\name{mushroom_dataset}
\alias{mushroom_dataset}
\title{Mushroom data set}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8124 rows and 23 columns.
}
\usage{
mushroom_dataset
}
\description{
Mushroom data set
}
\keyword{datasets}
