% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_datasets.R
\docType{data}
\name{my_dataset_list}
\alias{my_dataset_list}
\title{Data sets used in the Kohavi replication experiments}
\format{
An object of class \code{list} of length 6.
}
\usage{
my_dataset_list
}
\description{
Data sets used in the Kohavi replication experiments
}
\keyword{datasets}
