% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kohavi.R
\docType{data}
\name{my_fold_counts}
\alias{my_fold_counts}
\title{Fold counts used in the Kohavi replication experiments}
\format{
An object of class \code{numeric} of length 5.
}
\usage{
my_fold_counts
}
\description{
Fold counts used in the Kohavi replication experiments
}
\keyword{datasets}
