% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kohavi.R
\docType{data}
\name{my_model_list}
\alias{my_model_list}
\title{Models used in the Kohavi replication experiments}
\format{
An object of class \code{list} of length 7.
}
\usage{
my_model_list
}
\description{
Models used in the Kohavi replication experiments
}
\keyword{datasets}
