% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kNN.R
\name{nn_experiment}
\alias{nn_experiment}
\title{Investigate how quickly a k-nearest-neighbor classifier reaches capacity}
\usage{
nn_experiment(dims, be, reps = 50, seed = 2021)
}
\arguments{
\item{dims}{the number of dimensions used to determine the class}

\item{be}{the Bayes optimal classifier error}

\item{reps}{the number of times to repeat the experiment}

\item{seed}{random seed}
}
\value{
a list where the nth item is a vector of accuracies of models given
        n * 500 training examples
}
\description{
Investigate how quickly a k-nearest-neighbor classifier reaches capacity
}
