% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kNN.R
\name{run_nn_experiment}
\alias{run_nn_experiment}
\title{Repeat the k-nearest-neighbor experiment with various parameters}
\usage{
run_nn_experiment(dims = 1:3, bayes_errors = c(0, 0.1, 0.2), seed = 2021)
}
\description{
Repeat the k-nearest-neighbor experiment with various parameters
}
