% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_datasets.R
\name{simulate_data_type_1}
\alias{simulate_data_type_1}
\title{Simulate a dataset (type 1)}
\usage{
simulate_data_type_1(n, d = 1, bayes_error = 0.1)
}
\description{
These datasets have two classes, which are arranged in concentric shells.
Half of the columns are useless (no relationship with class).
}
