% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_datasets.R
\docType{data}
\name{soybean_dataset}
\alias{soybean_dataset}
\title{Soybean (large) data set}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 307 rows and 36 columns.
}
\usage{
soybean_dataset
}
\description{
Soybean (large) data set
}
\keyword{datasets}
