% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_datasets.R
\docType{data}
\name{vehicle_dataset}
\alias{vehicle_dataset}
\title{Vehicle silhouettes data set}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 846 rows and 19 columns.
}
\usage{
vehicle_dataset
}
\description{
Vehicle silhouettes data set
}
\keyword{datasets}
